refactoring embeds
