+++ tina/customFields/monaco-code-editor.tsx
@@ -1,6 +1,6 @@
-import React, { useEffect, useRef, useState } from "react";
-import { wrapFieldsWithMeta } from "tinacms";
 import debounce from "lodash/debounce";
+import { useEffect, useRef, useState } from "react";
+import { wrapFieldsWithMeta } from "tinacms";
 
 const MonacoCodeEditor = wrapFieldsWithMeta(({ input }) => {
   const editorRef = useRef<HTMLDivElement | null>(null);
